backend/requirements.txt
alembic==1.9.4
anyio==3.6.2
click==8.1.3
ecdsa==0.18.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
pyasn1==0.4.8
pydantic==1.10.5
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing-extensions==4.5.0
uvicorn==0.20.0
